+++ src/extension.ts
@@ -4,14 +4,15 @@ import createComponent from './createComponent';
 
 interface CreateComponentProps {
   args: any;
-  named: boolean;
+  named?: boolean;
   styled?: boolean;
   mobile?: boolean;
+  createNextPage?: boolean;
 }
 
-const handleCreateComponent = async ({ args, named, styled, mobile }: CreateComponentProps) => {
+const handleCreateComponent = async ({ args, named = false, styled = false, mobile = false, createNextPage = false }: CreateComponentProps) => {
   const componentName = await vscode.window.showInputBox({
-    prompt: `Enter the component name:`,
+    prompt: `Enter the ${ createNextPage ? 'page' : 'component' } name:`,
     ignoreFocusOut: true,
     valueSelection: [-1, -1]
   });
@@ -22,31 +23,34 @@ const handleCreateComponent = async ({ args, named, styled, mobile }: CreateComp
 
   if (args) {
     const path = args.fsPath;
-    createComponent(componentName, { dir: path, named, styled, mobile });
+    createComponent(componentName, { dir: path, named, styled, mobile, createNextPage });
   } else {
-    createComponent(componentName, { named, styled, mobile });
+    createComponent(componentName, { named, styled, mobile, createNextPage });
   }
 };
 
 export function activate(context: vscode.ExtensionContext) {
   let disposable = [
     vscode.commands.registerCommand("extension.create-react-component", args => {
-      handleCreateComponent({ args, named: false, styled: false, mobile: false });
+      handleCreateComponent({ args });
     }),
     vscode.commands.registerCommand("extension.create-react-styled-component", args => {
-      handleCreateComponent({ args, named: false, styled: true, mobile: false });
+      handleCreateComponent({ args, styled: true });
     }),
     vscode.commands.registerCommand("extension.create-react-named-component", args => {
-      handleCreateComponent({ args, named: true, styled: false, mobile: false });
+      handleCreateComponent({ args, named: true });
+    }),
+    vscode.commands.registerCommand("extension.create-next-page", args => {
+      handleCreateComponent({ args, named: true, createNextPage: true });
     }),
     vscode.commands.registerCommand("extension.create-react-native-component", args => {
-      handleCreateComponent({ args, named: false, styled: false, mobile: true });
+      handleCreateComponent({ args, mobile: true });
     }),
     vscode.commands.registerCommand("extension.create-react-native-styled-component", args => {
-      handleCreateComponent({ args, named: false, styled: true, mobile: true });
+      handleCreateComponent({ args, styled: true, mobile: true });
     }),
     vscode.commands.registerCommand("extension.create-react-native-named-component", args => {
-      handleCreateComponent({ args, named: true, styled: false, mobile: true });
+      handleCreateComponent({ args, named: true, mobile: true });
     })
   ];