+++ src/templates/typescript/reactNativeComponent.ts
@@ -1,13 +1,19 @@
 import CreateComponent from '../interfaces/CreateComponent';
 import creatReactImport from '../shared/functions/create-react-import';
+import pascalCase from '../shared/functions/pascal-case';
 
 export default ({ componentName, useReactImport }: CreateComponent) => (
-`${ creatReactImport(useReactImport) }import { View, Text, StyleSheet } from 'react-native';
+`${ creatReactImport(useReactImport, true) }import { View, Text, StyleSheet } from 'react-native';
 
-function ${ componentName }() {
+interface ${ pascalCase(componentName) }Props {
+  children: ReactNode;
+}
+
+function ${ pascalCase(componentName) }({ children }: ${ pascalCase(componentName) }Props) {
   return (
     <View style={ styles.container }>
-      <Text style={ styles.title }>${ componentName }</Text>
+      <Text style={ styles.title }>${ pascalCase(componentName) }</Text>
+      {children}
     </View>
   );
 }
@@ -26,6 +32,6 @@ const styles = StyleSheet.create({
   },
 });
 
-export default ${ componentName };
+export default ${ pascalCase(componentName) };
 `
 );